web/requirements.txt
Django==1.9.5
django-cors-headers==1.1.0
django-dotenv==1.4.1
django-filter==0.11.0
django-guardian==1.3.2
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
djoser==0.5.0
psycopg2==2.6.1
PyJWT==1.4.0
six==1.10.0
django-extensions==1.6.7
gunicorn==19.3.0
redis==2.10.3
requests